bidder

+++ models/bidder.go
@@ -26,7 +26,7 @@ type Position struct {
 
 type Cpm struct {
 	ID        int       `pg:"id,pk"`
-	AdID      int       `pg:"ad_id"`
+	AdID      int64     `pg:"ad_id"`
 	OldCpm    int       `pg:"old_cpm"`
 	NewCpm    int       `pg:"new_cpm"`
 	CreatedAt time.Time `pg:"created_at,default:now()"`
